Draw pages and their content in a scaled graphics context
Using a GraphicsContext, drawings can be scaled using wxWidgets.

But attention: Logic that changes the layout based on certain pixel constants might still use the old constants.

Click events on the canvas don't respect the scale factor yet.

And there is still a bug that text is not aligned correctly.