Respect zoom for calculations that take the available size into account
When calculating layout, we don't take zoom into account – instead, the zoom later gets added to the values that our layout code calculates.

When taking the available size into account, our layout code used the `GetClientSize` method of `MyCtrl`, which offered the real size of the rendered widget, which corresponds to the sizes *after* the zoom has been taken into account.

This caused things like the page not being centered when zoomed or the side-by-side layout mode refusing to show more pages when zooming out.

This commit introduces the `GetLogicalClientSize` method, which translates the actual client size to one that can be used when calculating the layout.